[kyo-ai] add Observe: turn observation and token usage tracking
Problem
-------
Every provider reports what a turn spent (tokens read, served from
cache, produced, spent reasoning) and kyo-ai discarded nearly all of
it: one backend kept a single subset field for one error message, the
others decoded nothing. Callers had no way to answer what a computation
cost, and no seam to react to turns at all: Mode intercepts at the
result tier (it sees the decoded value, never the wire reply, and does
not run on the streaming path), so passive concerns like usage
metering, transcript capture, or a spend guardrail had nowhere to
attach.

Solution
--------
Add Observe, the fifth enablement kind: the wire-tier, notification-only
counterpart of Mode. Where a mode receives the generation as a value and
may replace it, an observer receives each completed turn's
Completion.Reply and returns Unit, so it cannot change control flow; an
Abort in its capability row is a typed guardrail whose failure fails the
generation it fired in. Observers fire once per turn on the fiber that
ran it, at the moment the wire reply is read: before the ceiling
adjudication (a turn that aborts at the ceiling still spent its tokens),
on the generation and streaming paths alike, and inside forks, so a
losing race branch and a rolled-back forget block still report the turns
they completed.

All four backends decode their wire's usage into Reply.usage as AIStats:
input, cached-input, output, and reasoning-output tokens plus turns,
with the subsets kept Maybe so a reported zero stays distinct from a
wire that never breaks the number out. Observe.withStats collects stats
over a scope, alongside the result, with a per-instance breakdown
variant; Observe.init builds custom observers. Streaming carries usage
as stream elements the consumers sum, and a fully consumed stream fires
observers with the recorded synthetic result pair.

Notes
-----
- ai.stream previously ignored an instance's config override and
  enablements while ai.gen honored them; streamAgainst now applies the
  session env the way genLoop does.
- StopReason becomes public as part of Reply being the observer payload.
- The Claude Code kill-on-capture path reports a stated lower bound for
  output tokens (the terminal result event never arrives, so assistant
  events are summed after dedup by message id); both harness usage
  shapes were verified against live runs.

+++ kyo-ai/shared/src/main/scala/kyo/AIException.scala
@@ -161,7 +161,7 @@ case class AIOutputLimitException(
     // the number that picks the lever: a stop that spent nearly its whole allowance reasoning is not short
     // of ceiling, and raising the limit buys another expensive stop; one that barely reasoned is genuinely
     // short.
-    reasoningTokens: Maybe[Int] = Absent
+    reasoningTokens: Maybe[Long] = Absent
 )(using Frame)
     extends AIException(
         s"$provider stopped generating for model $model at the output-token ceiling" +
